Various improvements related to async tasks, saving, reloads, disabling.
* Improved: Catching a rare race-condition when trying to register a
task with the bukkit scheduler from within an async task while the
plugin is currently getting disabled. (see SchedulerUtils)

* Improved: Waiting (up to 10 seconds) for all currently running async
tasks to finish, before continuing to disable.

* Improved: Using synchronization to make sure that only one thread at
once attempts to write the save files. Previously this might have caused
issues, if an async and a sync save were going on at the same time (due
to the plugin being disabled).

* Changed: Added a method for requesting a delayed save. Refactored
(replaced) the chunkLoadSave task to use this instead now.

* Improved: If saving fails, it will now attempt another save after a
short delay.

* Improved: Renamed SaveInfo -> SaveResult and moved more saving related
variables into it. The debug output is now slightly more detailed as
well.

* Fixed: The SaveInfo / SaveResult object no longer gets reused. This
fixes an issue with invalid times being reported in case a sync save
gets triggered while an async save is already going on.

* A bunch of variables now get properly reset on reloads. Not sure if
this might have been an issue previously.