Improved / Fixed: Hiring behavior:
Changed: The hiring window can now always be open, regardless of if the player has the permission to hire this type of shopkeeper.
Added: Message 'msg-missing-hire-perm' if the player doesn't have the permission to hire shopkeepers at all.
Added: Message 'msg-cant-hire-shop-type' if the player cannot hire a certain type of shopkeeper (because he is missing the corresponding creation permissions).
Changed: Message 'msg-cant-hire' to use a different color.
Added: Config setting 'hire-require-creation-permission' (default: true) to allow enabling of hiring of shops the player would usually not be able to create himself.