During chunk unloads the plugin now checks and removes entities belonging to living entity shopkeepers located in other chunks.
This fixes an entity duplication issue (#488).

Additionally, the findOldEntity routine was slightly improved to also
find entities in neighboring (and loaded) chunks. (This is only useful
for situations in which chunk unload events are not properly called, or
world-save events were not properly called and the server crashes.
However, it only improves the handling of those cases, it doesn't fully
solve them).