More maven project restructuring:
* Removed reference to old Cube-Nation repository.
* Added a separate parent pom and renamed the old parent to 'sk-root'.
Only sk-main uses sk-root as parent now. The api project therefore
relies only on a minimal set of dependencies now.
* Shopkeepers and ShopkeepersAPI get installed and deployed with a
flattened pom now. This should allow depending on those without having
to also deploy the various versions of the parent poms or child modules.
* The different child modules that get bundled and deployed as part of
the final plugin jar use the 'provided' scope now, in order to not be
transitive for projects depending on Shopkeepers internals.
* Added a separate nms-parent pom, to further reduce duplicate contents
in the nms handler child modules.