Added 'data-version' to save data file.
This can be used to determine required migrations (not used for that
currently), or force a save of any loaded data (in the face of the
update to 1.13: Bukkit will convert ItemStacks during deserialization,
but we need to also save them back in the new serialized format to
complete the migration).

Any shopkeepers that cannot be loaded for some reason will also skip
this migration and will therefore have be migrated by hand. If this is
not done correctly, the save data may contain a mixture of legacy and
modern materials, which may cause for issues once Bukkit drops the
import for legacy materials at some point again in the future.

Other change included: Logging a message when a shopkeeper type ids gets
migrated.