Various changes to sign shopkeepers:
* Changed: Shopkeeper signs get despawned now with chunk unloads and
during world saves.
  * To prevent potential abuse of this temporarily despawning during
world saves, players are prevented from placing blocks at the location
where the sign is supposed to be. And shopkeeper signs will now replace
any block that is at the location the sign tries to spawn at.
  * Changed: Signs that could not be respawned will no longer cause the
shopkeeper to be deleted. Instead a 3 minute spawning delay was added to
prevent potential abuse (in case the spawned signs drop for some
reason).
  * Added a warning if the facing direction of a sign shopkeeper could
not be determined or is missing. Since the signs are getting removed
now, it will no longer attempt to determine the sign facing from the
world now. Instead it will fallback to using some arbitrary default
facing direction then.

* Changed: Blocks to which shopkeeper signs are attached to are now
protected as well.
* API/Internal: Spawn locations passed to the ShopCreationData are now
in the center of the spawn block and contain yaw and pitch to face the
creating player. Shop objects can now use this new information if they
wish.
* Added: Sign shopkeepers now support sign posts (if the creating player
targets the top of a block). The sign is rotated to face towards the
creating player.
* Added a setting 'enable-sign-post-shops' (default true) that can be
used to disable creation of sign post shops.