Various changes to chest protection.
* Fixed: Also preventing items being moved into a protected chest.
Players might have previously abused this to fill a chest with useless
items to prevent the shop from being able to trade due to being out of
storage space.
* Fixed: Item movement protection for double chests might not have
worked correctly.
* Added: The setting 'prevent-item-movement' (default: true) can now be
used to allow item movement from/to the shop chest. This gives the shop
owner more possibilities to manage the chest contents, but also opens
possibilities for other players to maliciously inject or extract items
if this is not properly prevented by other means. If chest protection is
disabled, this setting will have no effect (item movement will always be
allowed then).

* Reworked chest protection to by slightly less strict:
  * Players with bypass permission can now bypass block placement
restrictions when placing blocks next to protected shop chests.
  * Chests next to a shop's chest are now only protected / prevented
from being placed if they are actually connected to the shop's chest.
  * Placement of adjacent hoppers is now only prevented if they would
able to extract or inject items into the chest.
  * Added: Also restricting placement of droppers in a similar way now.

* Changed: If chest protection is disabled in the config, chest access
will not prevented either now. (It doesn't seem to make much sense to
prevent chest access but allow the chest to be broken/destroyed)
* Changed: If 'delete-shopkeeper-on-break-chest' is enabled, the
shopkeeper will now no longer be removed if the half of a double chest
gets broken that is not directly used by the shopkeeper.
* Changed: If 'delete-shopkeeper-on-break-chest' is enabled, the
shopkeepers will now also be removed if the chest gets destroyed by an
explosion. And if 'deleting-player-shop-returns-creation-item' is
enabled, shop creations items will be dropped for those shopkeepers as
well.