Slightly changed how the setup of the trading shopkeeper works.
Inside the editor, the player picks up items in their inventory and can
then freely place a copy of those items in the top window of the editor
to specify the traded items.

With this it is now also possible to setup multiple trades for the same
result item.

However, the number of trades that can be setup is strictly limited to 8
now. Previously one was able to setup additional trades by moving the
items in the chest around.

UIHandlers can now react to inventory dragging as well. The new trading
shopkeeper setup makes use of this to interpret dragging that only
involves a single slot just like clicking that slot (makes the inventory
interaction more fluent).