When a trade runs out of stock it gets deactivated, but is still visualized by limiting the trade's remaining uses.
The trades get updated for the trading player dynamically after every
trade attempt. This will not work in compatibility mode in which case it
will behave as before, with trading simply getting cancelled without
visual feedback for the user.

Also:
* Selling, buying and trading shopkeepers keep displaying their setup
trades in the editor, even if the corresponding items are no longer in
the chest.
  * Consequence: With this change it is no longer possible to work
around the limit of 8 trades per player shopkeeper by moving the items
in the chest around. This is left for a future change to improve.

Internal:
* UI handlers can now react to session starts and ends.
* Removed a few UI session cleanup calls that are redundant due to
bukkit calling corresponding inventory close events in those situations
which trigger cleanup as well.
* UI sessions keep track of the player directly now.