Changed book shop behavior to more closely match minecraft's behavior:
Only original and copies of the original book can be copied. When
copied, the book's generation is increased.

Also:
Trades for books that are no longer in the chest will be inactive but
still displayed in both the editor and the trading menu in the form of
dummy books with unknown author and 'tattered' generation. This is
similar to the new behavior of the other player shopkeepers.

Added message: msg-unknown-book-author