Unified trade representation among all types of shopkeepers to be more consistent and easier to remember:
The top item is now always representing the result item and the next two
items represent from bottom to top the buy items 1 and 2.

The only exception to this are the selling and book shopkeeper: Here the
high cost item will appear as buy item 2 inside the editor, but as first
item inside the trading recipe.

The admin and trading shop editors were updated to reflect those
changes. This change might initially confuse anyone who is used to the
previous admin trades setup (with the result item being at the bottom),
but should ultimately be more easy to remember due to being consistent
among all shopkeeper types.

Internal refactoring:
* Moved CommonEditorSetup into EditorHandler base class. The admin
shopkeeper makes use of this as well now.
* The CommonEditorSetup now also includes the common logic for
extracting the resulting trading offers from the editor.
* Moved UI handlers for the different shopkeeper types into separate
classes.
* Added global variables for some common inventory slot indices.