CommandArgument#parse now also returns the parsed value.
This is useful when there is a chain of wrapped arguments and the parent
needs to handle the value parsed by the child argument in some way.

Fixed: Chains of FirstOfArguments should now be able to properly store all the values parsed by child arguments along the chain. Previously this only worked for a chain depth of 1.