Changed the new argument fallback mechanic slightly:
Previously it would abort the parsing beyond the fallback after the
first command argument was able to successfully parse an argument from
the input. In some cases this is however not sufficient to determine
whether parsing past the fallback succeeds or not.
Example: Command "/list [player] [page]" with input "/list 123 2": '123'
is a rare but valid player name. We want to use the fallback mechanic as
heuristic, but still be able to correctly parse the input if both
arguments are specified explicitly.
Previously '123' would get interpreted as page number, causing the
overall parsing to fail to due an unparsed argument at the end.
The fallback mechanic was changed to now try to fully parse the input
beyond the fallback. If there are remaining unparsed arguments, the
parsing past the fallback is considered to have failed and will then
lead to the fallback to be used (accepting any input for the player
name), resulting in the above example to get parsed correctly again.

Other command related changes:
* Split up the Command#parseArguments function into separate parts.
* Replaced CommandArgs with ArgumentsReader: The arguments are now
stored inside the CommandInput and the ArgumentsReader only references
them from there.
* CommandContext is now an interface (with SimpleCommandContext as
implementation). A new class CommandContextView was added that gets used
anywhere where accessing the context is allowed, while modifying is not.
* BufferedCommandContext#getMapView() was fixed. This now creates a new
SimpleCommandContext with the combined values.
* Command and tab completion handling was moved from BaseCommand into
Command.
* Resetting of the arguments reader if parsing of an command argument
failed was moved from CommandArgument#parse into Command.
* Removed the lazy initialization of Command's argument list and
CommandContext's values map. ArrayList and LinkedHashMap are smart
enough to not actually allocate the internal array until it is required
for the first element.
* Removed Guava dependency for reverse iterating arguments in Command
class.
* Command#parseArguments will now check for unexpected input arguments
even if the command doesn't use any arguments.