Fixed: Inconsistent UI state when interacting with shopkeepers while holding a written book in hand.
Due to a Minecraft bug (MC-141494) interacting with a villager while holding a written book in the main or off hand results in weird glitches and tricks the plugin into thinking that the editor or trading UI got opened even though the book got opened instead.
We therefore ignore any interactions with shopkeeper mobs for now when the interacting player is holding a written book.