Fixed: Trading via shift-clicking while the player is charging a trident would allow the player to duplicate the trident.
The problem is that Inventory#setStorageContents updates all inventory
slots (even those that haven't actually changed) and Minecraft is not
able to deal with changes to the currently used ItemStack instance.

The fix is to only update those slots that have actually changed by
comparing the new inventory contents with the current ones before
updating individual slots. This also has the minor benefit of not having
to send the player inventory slot update packets for slots that haven't
actually changed.

This only works if the active item is not actually affected by the
inventory manipulation.