Fix: Cancel item interactions while interacting with a shopkeeper.
When a player interacts with a shopkeeper entity while holding an item
in hand, we may first receive the entity interaction event, which starts
an UI session, and then the interaction event for the item. To not start
any item actions for the held item (such as trident charging, food
consumption, etc.), we cancel any interaction events while an UI session
is already active.