The `/shopkeeper remote` command can be used from console now and optionally accepts a player as argument.
The shop will then be opened for the specified player.

Opening shops for other players requires the new 'shopkeeper.remote.otherplayers' permission (default: op).
The msg-command-description-remote message has been adapted to the addition of the optional player argument.

Internal: The description of the shopkeeper.remote permission has been updated to reflect that this commands also works for non-admin shops.