Added config option 'max-trades-pages' (default: 5, min: 1, max: 10).
This setting allows changing the number of pages that can be filled with
trading options.

This limit applies to all shopkeepers (there are no different settings
for different types of shops or different permission levels so far).

This allows for up to 90 trades per shopkeeper now. Note: This seems to
work fine. However, the scroll bar rendered by the
Minecraft client will only work nicely for up to 64 trades.