Drastically improved the performance of item conversions.
Instead of serializing and deserializing the items to/from a Yaml config, we only convert and reload the ItemMeta to/from its serialized form.
In my tests this has reduced the worst case timings from around 30+ ms to around 3.5 ms (still not good, but a lot better compared to before).

Other changes:
* Internal: Moved the code handling the serialization and deserialization of ItemMeta from ItemData into ItemUtils. This is used for our item conversions as well now.
* Debug: The item conversion debug messages include the timing information now.
* Debug: Since the item conversions always have some performance impact, even if no items got actually converted, we always print the debug messages now (if debug option 'item-conversions' is enabled) to allow debugging the item conversion times.