Moved the default conversion between Shopkeeper's TradingRecipe and Bukkit's MercahntRecipe into ShopkeeperUtils.
This can still be overridden inside subclasses of TradingHandler.

Also: The created villager trading recipes use a 'max-uses' limit of the maximum integer value now (instead of 10000). If the trade is 'out of stock' both the 'max-uses' and the 'uses' are set to 0. This change should probably not affect anyone though.