Fixed: Monster shopkeepers no longer burn when standing in sunlight.
Monsters usually get set on fire quite frequently when standing in sunlight. We therefore give zombies and skeletons a stone button as helmet now. This prevents them from getting set on fire in sunlight, by which we have to process less EntityCombustEvents. The stone button is also small enough to not be visible inside their head.

Entities standing in lava or fire are still set on fire as before.