Fixed: Enabled living shops specified inside the config would previously not get enabled if they were not perfectly matching the entity type name.
The enabled living shop types get validated and cached as part of the derived settings now.

Also: Removed the 1.16 'PIG_ZOMBIE' migration. We no longer automatically remove this mob type from the config, but only log a warning and then ignore it.

Also refactoring: Moved the validation of settings into a separate method.