Reverted: We ignore cancelled entity interaction events again.
This change has been problematic since we cannot differentiate between cases in which we can safely ignore the already cancelled event (eg. to bypass protection plugins) and cases in which we certainly do not want to execute our action. There can even be cases within our own plugin in which we only want one event handler to execute its action (for instance, both the shop creation listener and the sign shop listener react to interactions).

In order to nevertheless resolve conflicts with other plugins that react to interactions on LOWEST event priority as well (such as for example GriefPrevention), we now forcefully reorder the registered event handlers so that our event handlers are always executed first. Unlike the 'loadbefore' entry in plugin.yml file, this solution does not require to be aware of specific other plugins, and it does not break on dynamic reloads of the Shopkeepers plugin.

We log a debug notice whenever we move our event handler(s) in front of another plugin's event handler.