Improved handling of disabled shop (object) types.
We log a warning now whenever a shopkeeper is loaded or created that uses a shop type or shop object type that is disabled. We still load the shopkeeper, so that in the case of a player shop the container is still protected. However, even though the shopkeeper might in some cases seem to still work as normal, there is no guarantee for this. Admins are advised to either delete the shopkeeper, or change its shop object type to something else.

We also no longer spawn shopkeepers whose object type is disabled.