Further performance improvements to the shopkeeper mob AI.
Shopkeeper mobs with AI are now stored by chunk. Instead of iterating over all mobs and checking if their AI or gravity is active, we now iterate over the chunks first and skip those with neither active AI nor gravity.

Additionally, we skip the processing of mobs altogether if we know that there are no chunks with neither active AI nor gravity.