Changes to the initial offsets that are used to distribute ticking activities across multiple ticks.
These offsets are now (again) between 1 and the intended threshold that determines the rate at which the activity is supposed to take place. Previously, these offsets were added on top of the threshold, resulting in an initial delay before any of these activities started to take place (especially noticeable after plugin reloads). This change ensures that the thresholds not only determine the rate at which the activities take place, but also the maximum wait times until the activities take place.

The CyclicCounter has been extended with a lower bound so that it can now also start counting from a value different to 0.