The check for whether the shopkeeper mob moved and needs to be teleported back reuses the previous spawn location now.
Previously, this would freshly calculate the spawn location each time (which also involves a ray trace to find the exact location to stand on).

This change also ensures that shopkeeper mobs no longer start to fall if the block below them is broken and gravity is disabled. However, to account for shopkeepers that have previously been placed above passable or non-full blocks, which might also have been broken since then, we still spawn shopkeeper mobs up to one block below their location. So even though shopkeeper mobs will no longer immediately fall when the block below them is broken and gravity is disabled, they will still be able to spawn at the block below their location when they are respawned (i.e. when the chunk is reloaded, the plugin is reloaded, etc.).

Another side effect of this change is that if gravity is disabled, it is no longer as easily possible to force the shopkeeper mob to move back to its original location after the block it was previously standing on is broken. The shopkeeper will only check for an updated spawn location when it is being respawned now. If the mob is supposed to dynamically move when the block below it is broken, gravity needs to be enabled.