Config class fields are now made accessible only once right away.
These Field instances are copies and not shared with other components that potentially access these fields, so there is no need to restore their previous accessible state.

Marking these fields as accessible is also required when getting the value.

Also:
If getting a field's value fails, we throw a RuntimeException now instead of logging the error and then returning null.
And the exception message when setting the field's value fails has also been changed.