API: Renamed TradingOffer to TradeOffer.
This is more consistent with the naming of the other types of offers.
This is a breaking change. However, since this update already contains several other breaking API related changes, and the API is not yet declared to be stable anyways, this is not considered to be a major issue.