Added ItemUtils#parseMaterial.
This is similar to Bukkit's own Material#matchMaterial function, but allows for more flexibility on our side because we can make additions to this in the future. This is now used by us everywhere, including by Material settings.

Also slightly refactored the PlaceholderItems to prepare for placeholder items that derive their substituted item differently, other than looking up a Material by name.