Improved the inserted dummy trades when trades of a shopkeeper dynamically change.
When trades are removed while a player trades with a shopkeeper, we would previously insert empty dummy trades at the end of the trades list. This ensures that the number of trades never decreases while a player is trading, even if the shopkeeper's trades dynamically change in the meantime. This is required because the Minecraft client does not properly deal with the list of trades dynamically shrinking in size. However, these empty dummy trade are rather unusual and may therefore be confusing to players. Therefore, we now try to (heuristically) guess the trades that were removed and then insert blocked dummy trades that correspond to those.

This heuristic only works if the items of trades don't change, and if any new trades are only added at the end of the trades list. Otherwise the result may be sub-optimal, but still achieve the goal of preventing the list of trades to shrink in size. The benefit of the chosen heuristic is its simplicity and the relatively low computational overhead compared to other possible algorithms for determining more optimal matches between the old and new trade lists.