The utility methods for replacing ConfigurationSections with Maps use inner iterations now.
Also, we recursively replace all sections with Maps now before we update the current entry. This should have no effect, but makes it more clear that we set the entry to the deeply updated value. It also guards against the (unlikely) case that the top-level ConfigurationSection might eventually change to copy the Map that is passed to it, rather than inserting it directly.