Extended the ItemData tests.
The repair-cost meta was not tested before, as well as items with BlockStateTag.
However, there seems to be an open Spigot issue that affects items with BlockStateTag and causes them to not properly persist their data during deserialization currently (SPIGOT-6257). The corresponding new test cases therefore had to be disabled for now.