The TradeMerger merges trades over a longer duration now (up to 15 seconds), but still ensures that we wait a certain maximum duration (5 seconds) for a successive trade before we prematurely abort the merging.
The previous 5 seconds merge duration was unrealistic to cover typical trading sessions that involve trading multiple times for the same recipe when the player has to restock the item stacks in the input slots.

In order to avoid restarting the timeout task too often, we only restart it if the previous timeout task has not yet been running for longer than 0.5 seconds.