More changes to the next merge timeout task of the TradeMerger.
Instead of restarting this task up to once every 0.5 seconds, we let it run from start to finish and then adjust the task duration of the subsequent next merge timeout task according to when the last trade occurred.

We still avoid starting the task in cases where it does not make much of a noticeable difference. But the threshold for that is 0.3 instead of 0.5 seconds now.