In order to represent the logged trades more compactly, we merge equivalent trades that are triggered in quick succession over a certain period of time.
Added settings 'trade-log-merge-duration-ticks' and 'trade-log-next-merge-timeout-ticks', which control the maximum duration during which trades can be merged. They also allow the trade merging to be disabled.
By default, we merge successive equal trades for up to 15 seconds, but only if there is no gap larger than 5 seconds between the individual trades.

Other changes: Slightly reordered the item fields of TradeRecord to be consistent with the order in which they are logged by the CSV trade logger.