Added: Players now have to confirm when they delete a shopkeeper via the editor.
The confirmation is done by pressing a corresponding button in an inventory menu that opens. Canceling the action returns the player back to the editor. When the confirmation inventory is closed by some other mean, an abort message is printed.

If the shopkeeper is already removed by other means in the meantime, we send the 'shop-already-removed' message. We also send the 'shop-removed' message now when the shopkeeper is deleted via the editor menu.

Added messages:
* 'confirmation-ui-delete-shop-title'
* 'confirmation-ui-delete-shop-confirm-lore'
* 'confirmation-ui-confirm'
* 'confirmation-ui-cancel'
* 'confirmation-ui-cancel-lore'
* 'confirmation-ui-aborted'