Fixed: The living shop object types were not registered in the order specified inside the config.
This also fixes an error during plugin startup: The living shop object types were initialized before the Shopkeepers API and the unmodifiable item stack factory are bound. This resulted in an error because this initialization also triggered the initialization of the Settings class, which relies on the unmodifiable item stack factory.