Another attempt to get JitCI working.
JitCI doesn't seem to actually use the jitpack.yml, but needs to be configured online.
However, switching the active JDK version via the commands there is tricky (the commands sdk and source do not work).
Therefore: Added a separate 'build.sh' which installs and switches the active JDK version, and then triggers the Maven build from within this bash session.

The jitpack.yml is not expected to require this (but we will see about that).