Fixed an issue with Citizens player NPC shopkeeper nameplates being limited to a length of 16.
Since v2.13.0, when not using the `always-show-nameplates` setting (i.e. the default case), the nameplates of Citizens player NPC shopkeepers were limited to a length of 16 and would otherwise not be displayed. The issue is that the Citizens 'hover' nameplate option does not work well for player NPCs.

When upgrading and encountering issues with nameplates of previously created NPCs not showing, one has to either use the shopkeeper editor to rename the NPC at least once, or use the `/npc name` command to toggle the nameplate visibility.