Added setting `save-citizen-npcs-instantly`.
This setting is disabled by default. When enabled, we trigger a save of all Citizens NPC data whenever the Shopkeepers plugin modifies a Citizens NPC.
Saving the Citizens NPCs is quite a heavy operation. Since the Citizens API does not yet provide an API to trigger an asynchronous save, we trigger the asynchronous save by invoking the `/citizens save -a` command in the console. As a side effect, this will print command feedback in the console whenever the NPCs are saved this way.

Also:
* Slightly modified the description of the `save-instantly` setting in the default config.
* Although this has never been observed to be an issue, the Citizens shop object accounts now for the Citizens NPC creation to fail for unknown reasons.
* Removed some unused code related to the Citizens integration.
* The Citizens shopkeeper trait always uses the getNPC method now rather than accessing the npc field directly.