Fixed: Writing files (such as when saving the shopkeepers data) failed on some servers due to the SecurityManager implementation blindly denying the execute permission without differentiating between files and directories.
We check the execute (i.e. access) permission for directories in which we want to write files in order to provide more meaningful error messages in cases in which missing directory permissions prevent us from writing files. However, this intervention of the SecurityManager should have no effect on whether or not we can actually write to a directory, so we can safely ignore it.