Fixed: Turned another warning message related to shopkeeper spawning failing into a debug message.
There is an open Spigot 1.17 issue that sometimes causes the plugin to assume that it failed to spawn shopkeepers during chunk loads. In v2.13.1, we already turned a warning message related to this to a debug message in order to not spam the server log with warnings. But we missed this message.