Fixed: The 'style' property of horse shopkeepers was not getting saved anymore since v2.13.0.
This also caused migration warnings when the plugin attempted to load or save this property.

If you already upgraded, and don't have a backup of your save.yml file available, or don't want to revert to the backup's state, you may have to freshly edit the 'style' of your horse shopkeepers again.