Fixed: Some method calls would not be remapped when they are called through a subtype.
For example, the call nms.Mob#setYBodyRot(float) would not be remapped, whereas casting the mob to LivingEntity and then using nms.LivingEntity#setYBodyRot would successfully remap the call.

The issue was caused by trying to use both the '-jar' and '-cp' option at the same time when invoking SpecialSource, which results in the classpath option to be ignored.