Fixed: Shopkeepers could lose some of their trades during item migrations.
When an item was migrated, but the subsequent trades did not require any item migrations, these subsequent trades were lost.

This issue could for example be encountered on a Paper server when setting up trades with items that have multiple enchantments: During the shopkeeper's item migrations, Paper might reorder the item's enchantments, which the Shopkeepers plugin detects as an item migration, but would result in any subsequent trades of the shopkeeper that did not require any migrations to be lost.